+++ webpack.config.js
@@ -78,8 +78,7 @@ module.exports = {
     /** импорт файлов с расширением по умолчанию */
     extensions: ['.js', '.ts', '.css', '.scss'],
     alias: {
-      '@': path.resolve(__dirname, 'src'),
-      '~/*': path.resolve(__dirname, 'src'),
+      '~': path.resolve(__dirname, 'src'),
     },
   },
   devtool: isDev ? 'source-map' : '',
